Rows & Cols Bounds, Warning if Not All Words Fit, Font Size
Rows or cols could go below 1 which would cause the app to error out.
This commit fixes it by ignore values less than 1. Additionally, once
you got more than 50 rows or cols, performance got slow. This treats
all numbers higher than 50 as 50.

If a puzzle cannot fit all the words, it's helpful to know so you can
adjust the puzzle as needed. A warning now shows up if not all the
words could fit into the puzzle

Bumps up the font-size to help print quality so things aren't so fuzzy
when you save the screenshot and embed it in an external directory.